refactor: removes shelves state
It was inefficient to use an array as the source of truth for book shelves, passing that down through multiple components. This reverts to using book.shelf as the source of truth. To achieve this the book object has to be passed into the book component for the benefit of moving a book from the search results. This allows a book not currently in the state to be moved without a further (slow) API call.